+++ src/main/java/com/example/demo/domain/conversation/service/command/complete/ConversationCompleteCommandService.java
@@ -14,4 +14,9 @@ public interface ConversationCompleteCommandService {
      * 스토리 생성 진행 실패 -> 스토리 상태 업데이트
      */
     void failStory(Long storyId, Story.StoryStatus failedStatus);
+
+    /**
+     * 배치 대상 스토리의 retry_count를 +1하고 스토리 생성 이벤트 발행
+     */
+    void retryFailedStories(Long storyId, Long sessionId);
 }
