+++ src/main/java/com/example/demo/domain/story/service/command/StoryCommandService.java
@@ -7,9 +7,4 @@ public interface StoryCommandService {
      * @param storyId 스토리 ID
      */
     void markStoryVideoAsMaking(Long storyId);
-
-    /**
-     * 배치 대상 스토리의 retry_count를 +1하고 스토리 생성 이벤트 발행
-     */
-    void retryFailedStories(Long storyId, Long sessionId);
 }
